Incorporate addBubbles() into PBSmapping.R
Added addBubbles() to PBSmapping.R and made many changes including
completely reformatting all of the code, improving the comments,
changing some of the logic, and addressing bugs.  The bugs included
(a) testing floating-point equality using ==, which would result in
inconsistent behaviour and (b) sizing bubbles in a way that that was
inconsistent with the documentation.