Add refocusWorld() to help plot maps centred on the mid-Atlantic (issue #4)
This new function accepts a PolySet containing one or more polygons
with X-coordinates that collectively span approximately 360 degrees.
The function effectively joins the PolySet into a cylinder and
then splits it at an arbitrary longitude according to the
user-specified limits.  Modifications in the resulting PolySet
are restricted to shifting X-coordinates by +/- multiples of 360
degrees, and instead of clipping polygons, the return value simply
omits out-of-range polygons.