Improve long/lat <=> UTM conversions; update documentation (Issue #9)
Issue #9 noted erroneous results given data that spanned more than one
zone to the right/left of the intended central zone.  This update
improves our implementation by correcting a number of (implementation)
errors (described below), but the problem still persists.

The earlier (2001) implementation had a number of optimizations that
made it terribly difficult to verify the code against the published
algorithm and inspecting the implementation revealed some errors:
(1) the calculation of 'n' used scaled 'a'/'b' rather than original
    'a'/'b',
(2) the calculation of 'e^2' used scaled 'a'/'b' rather than original
    'a'/'b', and
(3) the calculation of 'm' was not iterative as suggested in the
    algorithm.
This new version more closely matches the published algorithm and is
consequently easy to verify.  The performance is similar to the old
algorithm (faster in some cases, slower in others).