Add support for S hemisphere to LL <=> UTM conversion (convUL)
Add support for the southern hemisphere to convUL (and supporting
functions) by adding a new "southern" argument.  By default, the
argument is NULL because detection is automatic when the source is LL.
For UTM source data, the default is FALSE because we typically work
within the northern hemisphere.

Add the ability to compile conversions.c (the supporting C code) as a
stand-alone application.  When executed, the stand-alone executable
performs a number of conversions that help validate the algorithm.