Support mult. bg colours, min.size arg, ... in addBubbles (issue 15)
Added support for multiple background colours to addBubbles.  Given a
vector of colours (length > 1), the function will now interpolate for
event$Z values within the range of the legend.  Added a min.size argument
to support visible bubbles even when event$Z is 0.  Enhanced behaviour
of legend.breaks argument so that the user can now specify the number
of breaks.  Fixed a bug where the zero label wouldn't be plotted for
legends in the "vert" layout.