Fix the spacing of xlab/ylab for the projection "LL"
In plots with the projection "LL", the code used "expression" to
produce the degree symbol for the axis labels. A side effect of this
approach caused the spacing to differ between the xlab and ylab. This
commit replaces "expression" with the Unicode character for the degree
symbol and corrects the spacing.